+++ spec/directions/bake_chapter_glossary/v1_spec.rb
@@ -78,6 +78,35 @@ RSpec.describe Kitchen::Directions::BakeChapterGlossary::V1 do
     )
   end
 
+  let(:chapter_with_para) do
+    chapter_element(
+      <<~HTML
+        <div data-type='glossary'>
+          <div>
+            <dl>
+              <dt>ZzZ</dt>
+              <dd>
+                <p>Test 1</p>
+              </dd>
+            </dl>
+            <dl>
+              <dt>ZzZ</dt>
+              <dd>
+                <p>Achoo</p>
+              </dd>
+            </dl>
+            <dl>
+              <dt>ABD</dt>
+              <dd>
+                <p>Test 2</p>
+              </dd>
+            </dl>
+          </div>
+        </div>
+      HTML
+    )
+  end
+
   context 'when append_to is nil' do
     it 'works' do
       metadata = metadata_element.append(child:
@@ -213,6 +242,49 @@ RSpec.describe Kitchen::Directions::BakeChapterGlossary::V1 do
           HTML
         )
       end
+    end
+  end
+
+  context 'when description has para' do
+    it 'works' do
+      metadata = metadata_element.append(child:
+        <<~HTML
+          <div data-type="random" id="subject">Random - should not be included</div>
+        HTML
+      )
+      expect(
+        described_class.new.bake(chapter: chapter_with_para, metadata_source: metadata, has_para: true)
+      ).to match_normalized_html(
+        <<~HTML
+          <div data-type="chapter">
+            <div class="os-eoc os-glossary-container" data-type="composite-page" data-uuid-key="glossary">
+              <h2 data-type="document-title">
+                <span class="os-text">Key Terms</span>
+              </h2>
+              <div data-type="metadata" style="display: none;">
+                <h1 data-type="document-title" itemprop="name">Key Terms</h1>
+                <div class="authors" id="authors_copy_1">Authors</div>
+                <div class="publishers" id="publishers_copy_1">Publishers</div>
+                <div class="print-style" id="print-style_copy_1">Print Style</div>
+                <div class="permissions" id="permissions_copy_1">Permissions</div>
+                <div data-type="subject" id="subject_copy_1">Subject</div>
+              </div>
+              <dl>
+                <dt>ABD</dt>
+                <dd>Test 2</dd>
+              </dl>
+              <dl>
+                <dt>ZzZ</dt>
+                <dd>Achoo</dd>
+              </dl>
+              <dl>
+                <dt>ZzZ</dt>
+                <dd>Test 1</dd>
+              </dl>
+            </div>
+          </div>
+        HTML
+      )
     end
   end
 end